[MetaModel] Rename Scalar_Staticness Control to retrigger CI
