+++ m2isar/metamodel/utils/scalar_staticness.py
